html / controller 업로드

+++ src/main/java/com/itpetshelter/itpetshelter/domain/Volunteer.java
@@ -13,9 +13,10 @@ import java.time.LocalDateTime;
 @Table(name = "Volunteer")
 @Entity
 public class Volunteer extends BaseEntity{
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int Vno;
+    private Long Vno;
 
     @ManyToOne
     @JoinColumn(name = "Sno")
